+++ changelog.md
@@ -3,6 +3,10 @@
 
 
 #### Unreleased (dev branch)
+- Double elimination. #9 - NicEastvillage
+    - Supports any number of teams and byes
+    - Supports bracket reset (an extra grand final match if the lower-bracket winner wins against upper-bracket winner)
+- Round robin and swiss now uses tiebreaker to rank all teams before transferring teams to the next stage. - NicEastvillage
 - Expanded, separated, and reworked functionality of ConfigFileEditor. #3 - cogitantium
 - Round robin and swiss now uses tiebreaker to rank all teams before transferring teams to the next stage. - NicEastvillage
 - Matches are now displayed slightly bigger. - NicEastvillage